Dev yizhen

+++ service/wrapper.R
@@ -5,7 +5,13 @@ library(limma)
 library(GEOquery)
 library(oligo)
 library(Biobase)
-
+library(rgl)
+library(Biobase)
+library(heatmaply)
+library(reshape2)
+library(plotly)
+library(reshape2)
+library(amap)
 
 
 process = function(){
@@ -84,6 +90,8 @@ process = function(){
           i<-i+1
           geneSet<-toString(args[i])
           i<-i+1
+          normal<-toString(args[i])
+          i<-i+1
           source<-toString(args[i])
            i<-i+1
           config_path<-toString(args[i])
@@ -98,15 +106,20 @@ process = function(){
                 celfiles = getCELfiles(projectId,listGroups,data_repo_path) 
              }else{
                 celfiles = getLocalGEOfiles(projectId,access_code,listGroups,data_repo_path) 
-             }
+          }
           saveRDS(celfiles, file = paste0(data_repo_path,"/celfiles.rds"))
-          norm_celfiles = QCnorm(celfiles,data_repo_path)
+          if(normal=="RMA")
+             norm_celfiles = RMA_QCnorm(celfiles,data_repo_path)
+           else{
+             norm_celfiles =loess_QCnorm(celfiles,data_repo_path)
+          }
+          # norm_celfiles = RMA_QCnorm(celfiles,data_repo_path)
           col_name<-pData(celfiles)$title
-          boxplot_DataBN<-list(col=col_name,data=t(norm_celfiles@listData[[3]]),color=pData(norm_celfiles[[11]])$colors)
-          RLE_data<-list(col=col_name,data=t(norm_celfiles@listData[[4]]),color=pData(norm_celfiles[[11]])$colors)
-          NUSE_data<-list(col=col_name,data=t(norm_celfiles@listData[[5]]),color=pData(norm_celfiles[[11]])$colors)
-          boxplot_DataAN<-list(col=col_name,data=t(norm_celfiles@listData[[8]]),color=pData(norm_celfiles[[11]])$colors)
-          tmp_pca<-norm_celfiles[[9]]
+          boxplot_DataBN<-list(col=col_name,data=t(norm_celfiles@listData[[2]][[1]]),ylable=norm_celfiles@listData[[2]][[2]],color=pData(norm_celfiles[[9]])$colors)
+          RLE_data<-list(col=col_name,data=t(norm_celfiles@listData[[3]][[1]]),ylable=norm_celfiles@listData[[3]][[2]],color=pData(norm_celfiles[[9]])$colors)
+          NUSE_data<-list(col=col_name,data=t(norm_celfiles@listData[[4]][[1]]),ylable=norm_celfiles@listData[[4]][[2]],color=pData(norm_celfiles[[9]])$colors)
+          boxplot_DataAN<-list(col=col_name,data=t(norm_celfiles@listData[[6]][[1]]),ylable=norm_celfiles@listData[[6]][[2]],color=pData(norm_celfiles[[9]])$colors)
+          tmp_pca<-norm_celfiles[[7]]
           pcaData<-list(
             col=colnames(tmp_pca$x[,1:3]),
             row=col_name,
@@ -116,20 +129,20 @@ process = function(){
             xlable=round(tmp_pca$sdev[1]^2/sum(tmp_pca$sdev^2)*100,2),
             ylable=round(tmp_pca$sdev[2]^2/sum(tmp_pca$sdev^2)*100,2),
             zlable=round(tmp_pca$sdev[3]^2/sum(tmp_pca$sdev^2)*100,2),
-            color=pData(norm_celfiles[[11]])$colors
+            color=pData(norm_celfiles[[9]])$colors
             )
 
           return_plot_data <- List(
-              norm_celfiles[[1]],
-              norm_celfiles[[2]]@listData,
+              "histBeforeNorm.html",
+              norm_celfiles[[1]]@listData,
               boxplot_DataBN,
               RLE_data,
               NUSE_data,
-              norm_celfiles[[6]],
-              norm_celfiles[[7]]@listData,
+              "histAfterRMAnorm.html",
+              norm_celfiles[[5]]@listData,
               boxplot_DataAN,
               pcaData,
-              norm_celfiles[[10]]
+              norm_celfiles[[8]]
               )
 
           #saveRDS(return_plot_data,file = paste0(data_repo_path,"/return_plot_data.rds"))
@@ -143,7 +156,7 @@ process = function(){
           # # or if using processCELfiles() function for test example, create this contrasts variable:
           # #cons = c("KO_1-Ctl_1","KO_2-Ctl_2")
        
-          diff_expr_genes = diffExprGenes(norm_celfiles[[11]],cons,projectId,data_repo_path)       #Call function
+          diff_expr_genes = diffExprGenes(norm_celfiles[[9]],cons,projectId,data_repo_path)       #Call function
           # # #### 4) l2p pathway analysis function, takes DEGs and species as input, returns list of up and downregulated pathways for each contrast ####
           # # # Output should dynamically respond to user-selected contrast
           saveRDS(diff_expr_genes, file = paste0(data_repo_path,"/diff_expr_genes.rds"))
@@ -166,19 +179,31 @@ process = function(){
           write(c(species,geneSet,config_path), paste0(data_repo_path,"/save_ssgsea_input.txt",sep=""))
 
            tryCatch({
-
-              ssGSEA_results = ssgseaPathways(diff_expr_genes,species,geneSet,data_repo_path,projectId,config_path)
-            
+              ss_result<-""
+              ssGSEA_results = ssgseaPathways(diff_expr_genes,species2,geneSet,data_repo_path,projectId,config_path)
+              message("ssgseaPathways finally")
+              file<-paste0(data_repo_path,"/",projectId,"_",cons,"_ssGSEA_pathways.txt")
+               if (!file.exists(file)) {
+                 ss_result<-read.table(file, header = FALSE, sep = "", dec = ".") 
+                 saveRDS(ss_result,file = paste0(data_repo_path,"/ssGSEA_results.rds"))
+                  
+              }
+             
+             
             },error =function(cond){
-                
+                 message(cond)
             },finally={
-                   message("ssgseaPathways finally")
-                   file<-paste0(data_repo_path,"/",projectId,"_",cons,"_ssGSEA_pathways.txt")
-                   ss_result<-read.table(file, header = FALSE, sep = "", dec = ".") 
-                   saveRDS(ss_result,file = paste0(data_repo_path,"/ssGSEA_results.rds"))
-                           re<-list(
-                    ss_name=names(ss_result),
-                    ss_data= ss_result[2:length(ss_result[,1]),],
+                     if(ss_result==""){
+                      ss_name_d= ""
+                      ss_data_d =""
+                     }else{
+                      ss_name_d=names(ss_result)
+                      ss_data_d=ss_result[2:length(ss_result[,1]),]
+                     }
+                    colorsD <-list(name=names(norm_celfiles[[10]]),colors=norm_celfiles[[10]])
+                    re<-list(
+                    ss_name=ss_name_d,
+                    ss_data= ss_data_d,
                     uppath=l2p_pathways[[1]][[1]],
                     downpath=l2p_pathways[[1]][[2]],
                     deg=diff_expr_genes$listDEGs[[1]],
@@ -198,7 +223,8 @@ process = function(){
                     group_2=cgroup2,
                     genSet=geneSet,
                     projectId=projectId,
-                    GSM=celfiles@phenoData@data
+                    GSM=celfiles@phenoData@data,
+                    colors=col2hex(celfiles@phenoData@data$colors)
                     )
 
                   write(toJSON(re),paste0(data_repo_path,"/result.txt",sep=""))
